+++ Core/SimpleTools/HideInWebGL.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+
+public class HideInWebGL : MonoBehaviour
+{
+#if UNITY_WEBGL
+    private void OnEnable()
+    {
+        this.gameObject.SetActive(false);
+    }
+#endif
+}