+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
+## [0.5.9] - 2026-04-21
+- Added SplashScreen and SplashScreenImage to GameFeatures
+- Added ControlSchemeManager to Input
+- Added PauseManager and PauseMenuController to Pause
+- Added RandomSprite to SimpleTools/2DHelpers
+- Added HideInWebGL to SimpleTools
+- Added SelectFirstSelectable to UI
+
 ## [0.5.8] - 2026-04-20
 - Added LayerMask helper ToLayerInt